Refactor doSourceMounts/newMountsCfg
